requirements.txt
requests==2.22.0
Pillow==6.2.2
python-dotenv==0.10.3
instabot==0.117.0